Decompose SortedIds to polynomial_base::{monomial_dyn, binary_ids}
