+++ rust/ommx/src/polynomial_base.rs
@@ -1,16 +1,20 @@
 mod add;
 mod approx;
 mod arbitrary;
+mod binary_ids;
 mod convert;
 mod degree;
 mod linear;
+mod monomial_dyn;
 mod mul;
 mod parse;
 mod polynomial;
 mod quadratic;
 
+pub use binary_ids::*;
 pub use degree::*;
 pub use linear::*;
+pub use monomial_dyn::*;
 pub use polynomial::*;
 pub use quadratic::*;
